[pymtl] changed clk vcd gen and line tracing (details)
I made two changes to the pymtl framework:

(1) I changed how VCD was being generated so that the VCD at least looks
right for the tests I am doing, and I also made the clock have a cycle
time of #10 instead of #2. This makes it easier to examine values during
the different clock phases in gtkwave and also makes the time in
nanoseconds directly correspond to the cycle number (if you divide by
10).

(2) I changed how line tracing works a bit. The print_line_trace function
will display the cycle number, and it doesn't need to be enabled. Instead
you just call this function on the simulator when you want to print line
tracing; this was necessary to print line tracing after we call
eval_combinational() but before we check the output ports. This allows
the line trace to show the state of everything _before_ we might cause an
assertion error and thus is more useful in debugging.