[pymtl] fix reg cb bug in sim & Bits zero test
The SimulationTool had a bug in cycle() where registers would
be added to the rnode_callback queue but never be removed. This
was a performance, but not functionality bug.

Certain tests were failing due to Bits objects storing a value
of 0 passing if conditions like so:

  if self.port.value:
    do_something()

Usually a value of zero is expected to fail this test, but by
default non-NoneType objects pass.  Defining the __nonzero__
function fixed this behavior.