[pymtl,pmlib,pex_*] fix issues with new bits/slicing (details)
I made some changes to Bits:

1. Added an assert if you try and create a Bits object with a value which
is too big to hold in the Bits object. This required removing some tests
in Bits_test.py and also a little extra work in some cases, but I feel
like a more explicit approach will avoid bugs where a user assigns a too
big of value to a Bits object but doesn't realize it is siliently being
truncated.

2. Added .int and .nbits accessors for Bits

3. Modified the width of the return type for Bits operators to correspond
to Verilog semantics.

4. Modified bit slices to return Bits objects as opposed to uints. You
can now copy a bits using [:]

5. Added a __str__ method which outputs the value of a Bits object in hex
at a fixed width corresponding to the bit width. This is useful for line
tracing. __repl__ is unchanged.

6. Added .sext() and .zext() methods to Bits

I added tests for these changes and then used them to fix issues in the
rest of the project. I was able to simplify line tracing significantly to
be something like this:

  def line_trace( self ):
    return "{} {} {} () {} {}" \
      .format( self.in0.value, self.in1.value, self.cin.value,
               self.out.value, self.cout.value )

The values will have the correct fixed width length because they know
their bitwidth.

I made some other changes which used to use workarounds, but now the
simple approach works.

I fixed issues related to most of the skipped tests that had to do with
not copying Bits; i.e., copying a pointer instead of the actual value. So
I added [:] in a couple of places to fix things.