[pmlib] TestSimpleMemoryNPorts commit (details)
Created a simple test memory parameterized based on the number of ports. I
created multiple port memory and the corresponding unit test file in a
seperate namespace to not conflict with the existing TestSimpleMemory. We
could rename the design later. The unit test for the multiple port memory
tests a single-port, dual-port and quad-port test memory. Line tracing as
well as all the tests including random delays pass.