[pmlib] SparseMemoryImage extn (details)
Extended the SparseMemoryImage class. We can now pass in a list of lists,
binary file name or an assembly string to create a sparse memory object.
Had to update the TestProcManager to use the renames in SMI. Also, added
two new customized options to py.test: --dump-asm and --dump-bin which can
dump the assembly file or the compiled binary for an assembly test. All
tests are passing.