[plab2] Integrated Muldiv to Stall Proc (details)
Fixed the division and remainder unit in Plab1. A lot of the bugs were due
to the difference in python division and remainder which differs compared
the that of the C/C++ language. Extended the muldiv unit to perform divu
and remu operations. TODO: NEED TO FIX ALL THE TESTS. THE TESTS ARE
CURRENTLY WRONG AND WILL FAIL.

Integrated the muldiv unit to the stalling processor and all tests pass.