[pymtl] fixed bug in transl reg detection
Register detection would sometimes mark input ports as registers if
their signals were written to in a parent modules logic block. In this
scenario the temporary wire used by the parent module should be type
reg, but obviously not the input port itself.

plab1_imul still not working, need to add parameterized module
definitions (ie, a register that takes nbits as a parameter).