[plab3] SRAMs model updated ( details )
Based on the discussion with Chris, I updated the SRAMs model to accurately
model synchronous and combinational SRAMs. I also found the IBM
applications note on Understanding Static RAM Operation quite useful,
although, it seems to be an old document. I updated the baseline design and
the incremental design. Everything is working. The state machines for both
the designs are the same with key differences being in the datapth.