[plab4] Cleaned up Crossbar + Possible bug? (details)
Crossbar model is a parameterized nxn square cross bar model. If we are
modeling this by having "n" number of output muxes for each output port,
then the select line is supposed to be log2(n). Previously select lines
were each a single bit wide, it passed all tests regardless. Maybe we need
a check for input assignments when in a port list? Maybe slices have a
corner case?