[pex_dac] added Shreesha's test harness (details)
Copied Shreeha's lab4 test harness to pex_dac, updated it to work with
RingRouter.  The RingRouters parameters were updated a bit to more
closely match those of Shreesha's router.

It looks like the router is working for certain individual tests, but
if too many messages are added to the TestSource it starts misrouting
packets for some reason.