[pex_dac] added bubble flowctrl to RingRouter (details)
The RingRouter now has bubble flowcontrol to prevent deadlock in the
RingRouter.  This passes for some of the deadlock tests in the test
harness, but fails for certain deadlock tests with delay.  Note that
increasing the buffering to 4 fixes this problem!  Although having
buffering of less than 4 prevents full-throughput, it seems like it
should still work...

Another thing to note is that a bug was found when connecting a constant
to a Wire.  This will be fixed in a future commit.  The current
workaround is to use a Port instead of a Wire for the signal you want to
connect to a constant.