[pymtl] basic physical elaboration (details)
Created a new PhysicalDimensions class for storing the dimensions of
bounding rectangles representing modules.  This is automatically
instantiated for each model during the elaboration phase, but does not
contain any information unless the model explicitly defines a
physical_elaboration() function.

It is up to each module to call physical_elaboration() on all its
children, add placement information to its children, and then use this
placement information to calculate and set its own height and width.

The Model base class adds a dump_physical_design() method which
recursively dumps the rectangles of all submodules in the design. This
information is only accurate if physical_design() has been called and
it's logic has accurately set child locations and its own dimensions!

A test example of a RegisterFile has been provided.  This is not
entirely accurate since each Register child cell is nbits instead of a
single bit.  However, it provides a basic, although artificial, example
of tiling in one dimension.