[v2pymtl] fixed verilator/sensitivity issue (details)
Verilated modules wrapped in PyMTL were not working with the new
sensitivity lists.  This is due to the fact that the PyMTL wrappers
generated by v2pymtl do not access .value, and therefore the
SensitivityVisitor would not pick these up as port reads.

The fix was to create an explicit "register_combinational" function that
takes a function name and a list of ports and uses this information to
manually create a sensitivity list.