[pymtl] fix portlist transl, found sense bug (details)
InPort and OutPort lists now translate correctly, this is done by
creating a temporary wire list whose indices are connected to their
flattened port counterparts.  To make this full useful we still need to
add translation support for for loops within @ blocks.

The test I created to verify portlist translation (DemuxNoLoop) found a
bug in the sensitivity list detection logic.  This bug occurs when a
signal is referred to with a .uint modifier like below:

  s.out[ s.sel.value.uint ] = s.in_.value

In this case the 'sel' signal is not added to the sensitivity list, most
likely due to a bug in the logic that does not detect this as a .value
access.  Need to fix this!