[pymtl] transl sep now _M_ instead of $ (details)
Previously signals that were connected to submodules would have a name
generated with the format submodule_name$signal_name.  Unfortunately,
our logic translation currently does not have a way to distinguish
submodule signals from ports that are part of a PortBundle.  As a
result, port names would have the format bundle_name$port_name, which
violates Verilog rules which disallow $ in port names.

Replacing the $ separator with _M_ fixed this problem, at the cost of
making the Verilog code considerably less readable.  Additionally,
PortBundles will translate but not Verilate properly since the v2pymtl
script needs to updated to special case the handling of ports with _M_
in the name properly.