[new_pymtl] reimpl slice-const connect (details)
The previous implementation of slice to constant connections required
that eval_combinational() be called before the constant value appeared
on the slices.  In addition to this be unintuitive, the implementation
created a callback that was placed on the event queue.  This is a bit
silly since the callback should only ever be called once (at
initialization) and never added to the event queue again (since it's a
constant value that should never be written).