[new_pymtl] created C++ EventQueue with cffi (details)
In hopes of improving the speed of the SimulationTool, I've implemented
an EventQueue module in C++ and interfaced it with Python via CFFI.

Only the module C++, Python interface, makefile, and Python unit tests
have been added.  The EventQueue module has not been integrated into the
SimulationTool yet.

No performance testing on the code has been done as of yet, so hopefully
it's actually faster than pure Python...

Also note: the C++ EventQueue is just a wrapper for the push()/pop()
methods of the STL queue+list, this means an allocate is probably
occuring for every enq()/deq().  A number of optimizations are possible:

* replace STL list with STL deque
* add support for "contains()", using a STL set or bitvector
* replace the STL queue with an STL list + hand managed head/tail ptrs
* do memory pool allocation:
  http://stackoverflow.com/a/1263122