[new_pymtl] bugfix SimTool port->bits swap! (details)
During simulator creation when instantiating the SimulationTool, we use
introspecton to find all the connected Signal types and reduce the many
nodes in a given connection to a single SignalValue type.  In addition
to reducing the amount of indirection, this provides our decorated logic
blocks with the preferred syntax.  However, in order for this to work,
we need to find all **names** pointing to Signal objects, and change
their assignment to that these same names instead point to SignalValue
objects.

Previously, this involved parsing the Signal .name attribute and using
this to regenerate the reference to the name.  Example:

  A signal referenced as "s.in_" in code would be swapped as:

   signal_obj.parent.__dict__[ 'in_' ] = signal_value_obj

  A signal referenced as "s.in_[ 3 ]" in code would be swapped as:

   signal_obj.parent.__dict__[ 'in_' ][ 3 ] = signal_value_obj

  A signal referenced as "s.in_.rdy" in code would be swapped as:

   signal_obj.parent.__dict__[ 'in_' ].__dict__[ 'rdy' ] = signal_value_obj

This nesting can get very complex and its really hard to recreate the
logic needed to access the name in this way. In particular this
became a problem once we started adding nested attributes in the form of
PortBundles and BitStructs, which allow arbitrarily complex attribute
access when nested.

There is a shortcut I found which is considered "unkosher" python, the
exec() function.  Instead of having complicated parsing and lookup
logic, we can just do the following:

  x = signal_obj
  exec( "x.parent.{} = signal_value".format( x.name ) ) in locals()

The locals() is necessary because a nested function is being used in
the context where we are calling exec, see:

  http://stackoverflow.com/a/4484946