[new_pymtl] fixed major bug in NetSinks! (details)
Previous implementation of the TestNetSinks did not check if a message
arrived redundantly, so it was possible for a test to pass even when it
should not.  The assertion was updated to catch this case.

In addition, an insideous bug was cropping up where message values were
not what they were expected to be.  This is because multiple Signal
objects were being passed the same copy of the msg_type.  This worked in
the past because msg_type was generally an int that was used internally
to construct **new** Bits objects for each Signal.  However, now that
msg_type can be a BitStruct which is an object (not a true type),
passing the same BitStruct to multiple ports will give them references
to the same object.  As a result, writing one will update the others!

The current work around is to always make a copy of the msg_type object
before passing it to signals or child modules as parameters. This is
very inefficient and confusing, and what we should do is make
instantiating a BitStruct return a type, and then have each Signal
instantiate that type internally to make sure a unique copy belongs to
each signal.