[new_pymtl] undo copy bugfix, fix in Signal (details)
The previous fix to the msg_type copy probablem wasn't scalable or user
friendly, since it required users to copy objects in their own code.

Instead every instantiation of Signal creates a deepcopy of the msg_type
passed to it.  Note that this **must** be a deepcopy, a shallow copy
caused some tests to fail due to some monkeypatching that is done in the
SimulatorTool (specifically adding to ._ucb).  Still not entirely sure
what the problem was, but deepcopy fixed it.