[proc] ParcProcPipelinedMul impl, working with asm tests (details)
The pipelined multiplier unit is now treated as a separate pipeline with
stages Y0, Y1, Y2, Y3, where Y3 writes back to the 2nd port of the
register file.  Note that this operates in parallel to the traditional
pipeline stages of X, M, W.  This implements the equivalent of an
I2O2 processor configuration which is okay since we don't have
exceptions currently.  The pipeline looks similar to:

  F D X  M  W
      Y0 Y1 Y2 Y3

Some caveats:

- Only mult operations use the Y pipeline, NOT the div/rem operations.
  This is a bit strange because mult/div/rem share live in the same
  datapath unit (IntMulDiv), but are treated like separate pipelines!

- Issuing multiply and divide operations back to back has not been
  tested, I have no idea if this works.

- I have not tested the corner case where a load operation issues before
  a multiply, but the load misses in the data cache.  This situation
  could cause a write-after-write hazard!

  Typical case:

    ld:  F D X M  W
    mul:   F D Y0 Y1 Y2 Y3

  Cache miss case (write-after-write hazard!):

    ld:  F D X .  .  .  M  W
    mul:   F D Y0 Y1 Y2 Y3