[pisa,pmlib_extra,proc_fl,mvmult_fl] see details
Refactored common classes like Bytes and the proxy classes into
pmlib_extra. We should eventually merge this into new_pmlib?

This commit includes support for a functional-level model of the
accelerator. We can use MatrixVec directly in the ISA simulator, but we
also have a wrapped version in MatrixVecFL which provides a port-based
interface. The MatrixVecProxy class can substitute for MatrixVec so that
the ISA simulator can appear to use the same interface, but really it is
proxied into a port-based interface. This way ParcProcFL can be connected
to MatrixVecFL.