[pisa] hacky bugfix: globals in jal/jalr tests (details)
Currently the test generator for jal/jalr bypassing tests uses global
state in order to enable creation of a single assembly file containing
multiple jump tests.  The state is used to calculate the link address as
well as the create unique labels.  The lack of a reset condition meant
that if the functions we're called multiple times (which they were for
various processor config tests), the link address would be incorrect and
the tests would fail.