[dotprod] hack: delete bytearray to reduce memory footprint (details)
Currently py.test observes huge memory bloat (3GB+ for the Tile tests)
due to the failure to free test objects: Model, SimulatorTool, or
at least many of the children the point to.

After lots of memory profiling, the vast majority of memory (80%+) was due
to the byte array which is used to emulate the physical memory.  The
current temporary solution is to simply zero out the contents of the
bytearray after the simulation completes.