[pymtl.tools.sim] fix for issue #79 (details)
The PyMTL SimulationTool uses five different mechanisms for propagating
updates between SignalValue (e.g. Bits) objects. Relationships between
SignalValues are expressed inside PyMTL models using Signal objects
(InPorts, OutPorts, and Wires):

  1. Tick Blocks          (notify_sim_sequential_update)
  2. Combinational Blocks (notify_sim_comb_update)
  3. Signal Connections   (SignalValue pointers)
  4. Sliced Connections   (_SignalSlice, slice_cb)
  5. Sliced SignalValues  (BitSlice)

Bug #79 was caused by an error in the interaction of Type 4 and Type 5
updates.

==========
Background
==========

1. Tick Blocks
--------------

Tick blocks are created from code in the following format:

  @s.tick
  def sequential_logic():
    s.out.next = s.in_

Tick functions are placed on a tick queue that fires every clock cycle
regardless of whether input signals have changed or not.  The write to
s.out.next is **buffered** and placed on a register queue, as a result the
value update is not immediately visible via reads of s.out. After all
tick blocks on the tick queue have been executed, any buffered values on
the register queue are applied by writing the signal's .value attribute
(in this case, s.out.value).

2. Combinational Blocks
-----------------------

Combinational blocks are created from code in the following format:

  @s.combinational
  def combinational_logic():
    s.out.value = s.in_

Sequential functions are placed on an event queue to be executed **only**
when one or more of their input Signals (in this case, s.in_) have
changed. The write to s.out.value is immediately visible, and can
potentially add other combinational blocks to the event queue.

3. Signal Connections
--------------

Connections are created from structually composing two Signals:

  s.connect( s.in_, s.out )

In the simulator, connections are implemented by collapsing all
interconnected Signal objects into a single SignalValue pointer.
Writing one Signal will immediately update all other interconnected
Signals (e.g., a write to s.in_.value with be immediately visible from a
read of s.out.value, and vice versa).

4. Sliced Connections
---------------------

Sliced connections are created from structurally composing two Signals,
where one or both Signals have been sliced:

  s.connect( s.in[0:4], s.a      )
  s.connect( s.in[0:4], s.b[4:8] )
  s.connect( s.in,      s.c[0:8] )

Slicing of a Signal (InPort, OutPort, or Wire) creates a _SignalSlice
object which is stored by the s.connect operator as a ConnectionEdge.
During simulator construction the information encoded by the
ConnectionEdge (source, dest, and slicing info) is used to create closure
functions (slice_cb). These slice_cb functions update destination
SignalValues from source SignalValues by performing any necessary slicing:

  dest[ dest_slice ].value = src[ src_slice ]

The slice_cb closures are attached to each source SignalValue (in the
._slices attribute), any write to a SignalValue's .value attribute will
cause all slice_cb closures in ._slices to be executed immediately. This
improves performance and observability of updates during debugging.

5. Sliced SignalValues
----------------------

Sliced SignalValues are created by writing a slice within a Tick Block or
a Combinational Block:

  @s.combinational
  def combinational_logic():
    s.out[0:4].value = s.in_

Slicing of a SignalValue must return another Bits object in order to
enable chaining of operations:

  s.a[0:4] + s.b[4:8]

However, in order to enable updating of s.out from a new Bits object it
returns via slicing, special BitSlice objects are used instead. BitSlices
are just like Bits but have information that point them to the original
Bits object they were created from and allow them to update that object
when .value is assigned to.  This information includes a reference to the
original Bit's slice_cb closures stored in ._slice.

===========
The Problem
===========

The bug came about because of two optimizations used in the creation of
SlicedConnections and Sliced SignalValues. In SlicedConnections we create
a slice_cb in the following way:

  src_addr  = c.src_slice  if c.src_slice  != None else slice( None )
  dest_bits = dest[ c.dest_slice ] if c.dest_slice != None else dest
  def slice_cb():
    # We need to slice the src each time.  This is because writing
    # to a BitSlice will updates the Bits it was sliced from, but
    # not vice versa.
    dest_bits.v = src[ src_addr ]
  return slice_cb

Note that dest_bits will contain a Bits if the destination is not sliced,
and will contain a BitSlice if the destination is sliced. This BitSlice
object is closed over and does not change or disappear; this optimization
is performed to prevent the need to create a new BitSlice on each write!

However, the construction of BitSlices formerly contained the following
code:

  self._slices = self._target_bits._slices

Where self._target_bits was a reference to the original Bits object the
BitSlice was created from. Alternatively, the BitSlice object could look
at ._target_bits each time it's _slices were referenced, but this
optimization prevented what seemed to be an unnecessary reference. This
was bad!

Given the following scenario:

  s.connect( x.a,    y.b[0:2] )
  s.connect( y.b[0], z.c      )

We chould be creating two slice_cb closures:

  dest1 = y.b[ 0:2 ]
  def closure_1():
    dest1.value = x.a

  dest2 = z.c
  def closure_2():
    dest2.value = y.b[0]

Note that closure_2 should be attached to y.b's _slices, so that anytime
y.b is updated it fires. However, if closure_1 is created before
closure_2, then the following line in BitSlices:

  self._slices = self._target_bits._slices

Will mean that dest1 will contain a BitSlice that **does not contain**
closure_2, and therefore writes to y.b will not properly update z.c! If
closure_2 is created before closure_1, everything will work fine.

============
The Solution
============

To fix this problem, we either need to a) not use a cached copy of the
BitSlice in the slice_cb, or b) not use a cached copy of ._slices in the
BitSlice object. I went with fix b) by making ._slices a property of
BitSlices.