[accel.dotprod] remove numpy dependency, fix issue with list mem adapter
We really don't want to create a very heavy weight dependency on numpy
just to show how we can use the numpy.dot function. We can just use our
own little dot function. After removing this dependency, I realized that
an earlier change to the ListMemPortAdapter broke the dot product
accelerator. For now the easiest thing to do was simply to add a
ListMemPortAdapterOld that works with the way the dot product accelerator
uses ReqRespBundles. These bundles may have been a mistake because we
don't have nested port bundles, so it prevents being able to grab a
ValRdyBundle from the ReqRespBundle which seriously undermines
composability with other parts of the framework.