[pymtl.tools.transl] workaround for issue #129
Fixes bug where BitStructs would not be marked as a reg in generated
Verilog if the attributes of the signal were written to in a behavioral
block. While this did not affect Verilator simulation, it caused problems
with Synopsys Design Compiler.

There is a caveat to this fix: if **any** field of a BitStruct is assigned
to in a behavioral block, the **entire** BitStruct is assumed to be a reg.

This means scenarios like the following could potentially be a problem::

  class MyModel( Model ):
    def __init__( s ):
       s.in_ = InPort ( Point() )
       s.out = InPort ( Point() )

    # structural connection of one field

    s.connect( s.in_.x, s.out.x )

    # behavioral assignment for another field

    @s.combinational
    def logic()
      s.out.y.value = s.in_.y

This can be viewed as consistent with our current lack of support for
using slices to treat one part of a signal as a combinational wire and
another part as a sequential register.