[proc/parc_fl] use new test mem, adapters, mtxcel/mfxcel (details)
I am not 100% sure the proc-fl-sim works. We really need to clean
up the bootstrap mess. Ideally, we need a way to test the processor
simulators w/o the cross-compiler. So we could eventually use the
pydgin "mini-assembler" to assemble some ELF binaries and then run
these on the processor simulators for testing purposes.