[pymtl/model] avoid hash collision for Verilog translation (see more)
This commit adds the module's __class__ to the hash generation to avoid
a hash collision that can break Verilog functionality.

When we generate a model's class name (used in Verilog translation), the
hash is currently generated based on the list of arguments and their
values. That means if there is the same class name and same args and arg
values (e.g., two Mux's each with 2 ports and 47 bits, but one is one-hot
and one is not), the hashes will collide. In Verilog, collided names
result in both modules pointing at the same module definition, so one is
incorrect. This commit prevents this kind of collision.