[pymtl.model] fix issue with collecting args for Models (details)
There was a bug related to collecting arguments for a Model when the
constructor was defined without default values and the Model was
instantiated using keywords. The following illustrates the problem:

 >>> from pymtl import *
 >>> class A(Model):
 ...   def __init__( s, arg1, arg2 ):
 ...     s.arg1 = arg1
 ...     s.arg2 = arg2
 ...
 >>> a1 = A( arg1=1, arg2=2 )
 >>> a2 = A( arg1=3, arg2=24)
 >>> a1.elaborate()
 >>> a2.elaborate()
 >>> a1.class_name
 'A_0x395645cdf5e6932f'
 >>> a2.class_name
 'A_0x395645cdf5e6932f'

These should have different class names because they have different
arguments. I have added a test case which failed due to the above issue,
I fixed the __call__ method in the metaclasses.py magic, and now these
test cases pass.