[pymtl/model] enable setting the class_name explicitly
If the parameters to a model are types, then apparently the suffix hash
changes everytime we run Python. This can be incredibly frustrating when
the top-level model we want to push through the ASIC flow is
parameterized in this way, since it means the suffix hash is always
changing and we need to constantly update the Makefrag. I added support
for a model to set a special member variable named explicit_modulename.
If this is present, then PyMTL will just use the given explicit module
name as the class name instead of trying to create a hash.