[pclib/rtl] add test for const_zero (details)
If we set the const_zero=True flag, then reading r0 should always result
in a zero, but it looks like this is not the case. This test will always
pass in PyMTL because PyMTL currently always initializes all signals to
zero, and it will also pass by default with --test-verilog. We can force
a failure by using --test-verilog='ones' which causes Verilator to
set all uninitialized state to all ones:

  % py.test ./RegisterFile_test -k const0
  % py.test ./RegisterFile_test -k const0 --test-verilog
  % py.test ./RegisterFile_test -k const0 --test-verilog='ones'