[pymtl/tools/integration] Create one simcontext for each model
This approach basically enables us to separate each test case.

If we keep a single systemc simcontext, and rely on "destroy" method
and the never-been-called "__del__" method to reset the systemc
sim context, then when a test case failed (usually an assertion),
the destroy method won't be called. As a result, the next test
case is not able to instantiate and will abort the whole py.test.

Now I create one sc_simcontext for each model. The above issue is
resolved by leaving that test case in memory just like what we
always do for verilated models. It increases a little bit memory
consumption (roughly ~5MB per undestroyed context).