[pytest/Bits] Adding slice assignment bounds check test cases
The cases show the problem in current implementation when a value is assigned to a slice of object Bits. Currently, the implementation relies on bit_length() provided by python. However, bt_length() is confusin and produces wrong bounds check. The problem does not happen during the creation of the Bits object as the bounds are checked using explicit _max and _min rather than number of bits. However, a wrong number of bits is usually reported in the error message, which is confusing.