[Bits/Bits_test] Changing the way minimum number of bits are calculated
This would eliminate the bugs happening in both:
the message displayed when the created Bits has an initial value that would overflow, and when a slice is assigned a value.

The assumption is that if the value is positive, we treat it as unsigned; however, if the value is negative, we treat it as 2's complement signed.

The old code relies on bit_length() provided by python int class. But, that is confusing as it does not include sign bit. Nevertheless, when the number is a negative exponentiation of 2, the sign bit is added to the return bit. To fix this, we either handle perfect negative 2's exponentiation differently, or we write our own code. In my opinion, handling the perfect exponentiation differently is rather dirty. Therefore, I wrote our own method to handle this.

Also, when Bits is instantiated with one bit, then there is no problem whatsoever in having -1 as an acceptable assignment. -1 can fit in one bit. We handle this exactly like the aforementioned rule. Negative values will be treated as 2's complement and thus 1 bit is enough for -1 to be represented.