fix issue with verilator and line tracing
We have to use some pretty complicated code to make Verilog line
tracing work with PyMTL. One thing we need is a way to set the scope
in our C wrapper so we can actually call the DPI line_trace function.
We do this using the svSetScope function:

  https://www.veripool.org/projects/verilator/wiki/Manual-verilator#DPI-System-Task-Functions

I tried using something like this:

  svSetScope( svGetScopeFromName("TOP.v.verilog_module") );

But I couldn't really get that to work, so ultimately I went with
this:

  svSetScope( &model->__VlSymsp->__Vscope_v__verilog_module );

The extra .v was because Verilator would add that mysterious level
of the hiearchy. However, it looks like this issue suggests that
this extra level has been removed:

  https://www.veripool.org/issues/1050-Verilator-Extra-v-in-full-signal-pathnames

So now I am trying this:

  svSetScope( &model->__VlSymsp->__Vscope_{model_name}__verilog_module );

where {model_name} is the actual Verilog module name which looks
like what Verilator uses as the instance name.