fix explicit_modulename with Verilog import
Verilator used to always name the top-level C++ harness "TOP" in the VCD
file, but it would also insert a top-level wrapper named "v". So this
meant TOP.v would refer to the module we were pushing through the flow,
and as a consequence we could just use TOP/v when specifying the source
instance for SAIF mapping to Synopsys DC and for power analysis to
Synoptsys PT.

  https://www.veripool.org/issues/1050-Verilator-Extra-v-in-full-signal-pathnames

Verilator made a recent change where they got rid of that wrapper and
instead in the VCD file you will now see something like TOP.ModuleName.
Essentially, Verilator is using the top-level modulename as the instance
name in the test harness. So now we use TOP/ModuleName when specifying
ghte source instance in Synopsys DC and PT. Unfortunately, this kind of
broke when combined with explicit_modulename.

The problem is that when we do Verilog import, PyMTL will use the
explicit_modulename but when we do translation the explicit_modulename
attribtue is no longer present so PyMTL will revert to using the hash.
This has for a long time caused odd behavior where Verilog RTL flows
would have -two- verilog files in their build directory. We always push
the one with the explicit modulename through the flow, but it turns out
the one actually being simulated is the one with the hash ... which means
the VCD file has the module name with the hash in it ... which means the
VCD/SAIF file no longer matches the actual Verilog we are pushing through
the flow.

The fix is just to make sure we copy over the explicit_modulename when we
use metaclasses to create the new model in the Verilog import tool. I
added a sime test that illustrated the problem.

In this process, I also saw .verilator2.vcd files starting to pop up
again. We used to use this:

  verilator_vcd_file = '{{}}.verilator{{}}{{}}'.format(filen, s.id_, ext)

but I am not sure why we neded s.id_. It was added in this commit:

 https://github.com/cornell-brg/pymtl/commit/c3de1d12

so something about generating separate files for each instance. But these
instances should all have different VCD file names to start off with, so
maybe this was before we setup our tests to ensure unique VCD file names?
The problem is it means that we end up with something like
.verilator2.vcd when we import Verilog RTL and .verilator.vcd when we use
a pure PyMTL model ... and so that means we need to special case things
for the ASIC flow. So for now I am removing the s.id_.